Add port and timezone to the server
The server program now needs the port to be passed as an argument using
the `p` flag. If not port is passed, the program won't run.
It also set the timezone in which the server with run using the
environment variable `TZ`, if the variable is empty or is an invalid
timezone the program use UTC instead.